+++ tests/spec/Gnugat/Redaktilo/FilesystemSpec.php
@@ -33,22 +33,6 @@ class FilesystemSpec extends ObjectBehavior
         $this->beConstructedWith($symfonyFilesystem);
     }
 
-    function it_detects_file_line_break()
-    {
-        $lines = array(
-            'King Arthur: One, two, five!',
-            'Sir Galahad: Three sir!',
-            'King Arthur: THREE!',
-        );
-        $fileWithoutLines = $lines[0];
-        $windowsFile = implode(Filesystem::LINE_BREAK_WINDOWS, $lines);
-        $otherFile = implode(Filesystem::LINE_BREAK_OTHER, $lines);
-
-        $this->detectLineBreak($fileWithoutLines)->shouldBe(PHP_EOL);
-        $this->detectLineBreak($windowsFile)->shouldBe(Filesystem::LINE_BREAK_WINDOWS);
-        $this->detectLineBreak($otherFile)->shouldBe(Filesystem::LINE_BREAK_OTHER);
-    }
-
     function it_opens_existing_files()
     {
         $this->fileCopier->copy($this->sourceFilename, $this->copyFilename, true);